agents/langgraph/agentic_rag/pyproject.toml
[project]
name = "agentic_rag"
version = "0.2.0"
description = "A LangGraph Agentic RAG app using OpenAI-compatible APIs. Supports retrieval-augmented generation with customizable vector stores."
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.12, <3.14"
dependencies = [
    "fastapi>=0.132.0",
    "uvicorn[standard]>=0.41.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-community>=0.4.1",
    "langchain-milvus==0.3.3",
    "langgraph-prebuilt>=1.0.0",
    "llama-stack>=0.5.0",
    "llama-stack-client>=0.5.0",
    "openai>=2.24.0",
    "pymilvus==2.6.9", # <-- 2.6.10 has a bug https://github.com/milvus-io/pymilvus/issues/3317
    "milvus-lite>=2.5.1",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0,<82.0.0",
    "chardet==7.2.0",
    "pypdf==6.9.1",
    "flask>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
]
tracing = [
    "mlflow>=3.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"
